requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.3
django-ckeditor==6.7.1
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.3
gunicorn==22.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kombu==5.3.1
packaging==23.1
phonenumbers==8.13.17
Pillow==10.0.0
prompt-toolkit==3.0.39
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-telegram-bot==20.5
pytils==0.4.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
